+++ usvg/src/convert/text/convert.rs
@@ -24,6 +24,7 @@ use crate::convert::{
     style,
     units,
 };
+use super::TextNode;
 
 
 /// A read-only text index in bytes.
@@ -42,13 +43,13 @@ impl ByteIndex {
     }
 
     /// Converts byte position into a code point position.
-    pub fn code_point_from(&self, text: &str) -> usize {
+    pub fn code_point_at(&self, text: &str) -> usize {
         text.char_indices().take_while(|(i, _)| *i != self.0).count()
     }
 
     /// Converts byte position into a character.
-    pub fn char_from(&self, text: &str) -> Option<char> {
-        text[self.0..].chars().next()
+    pub fn char_from(&self, text: &str) -> char {
+        text[self.0..].chars().next().unwrap()
     }
 }
 
@@ -93,6 +94,10 @@ impl FontData {
         self.descent as f64 * self.scale(font_size)
     }
 
+    pub fn height(&self, font_size: f64) -> f64 {
+        self.ascent(font_size) - self.descent(font_size)
+    }
+
     pub fn x_height(&self, font_size: f64) -> f64 {
         self.x_height as f64 * self.scale(font_size)
     }
@@ -113,6 +118,23 @@ impl std::fmt::Debug for FontData {
 }
 
 
+pub struct TextPath {
+    /// A text offset in SVG coordinates.
+    ///
+    /// Percentage values already resolved.
+    pub start_offset: f64,
+
+    pub segments: Vec<tree::PathSegment>,
+}
+
+
+#[derive(Clone)]
+pub enum TextFlow {
+    Horizontal,
+    Path(Rc<TextPath>),
+}
+
+
 /// A text chunk.
 ///
 /// Text alignment and BIDI reordering can be done only inside a text chunk.
@@ -121,6 +143,7 @@ pub struct TextChunk {
     pub y: Option<f64>,
     pub anchor: TextAnchor,
     pub spans: Vec<TextSpan>,
+    pub text_flow: TextFlow,
     pub text: String,
 }
 
@@ -167,20 +190,80 @@ pub enum WritingMode {
 }
 
 
+struct IterState {
+    chars_count: usize,
+    chunk_bytes_count: usize,
+    split_chunk: bool,
+    text_flow: TextFlow,
+    chunks: Vec<TextChunk>,
+}
+
 pub fn collect_text_chunks(
-    text_elem: &svgdom::Node,
+    text_node: &TextNode,
     pos_list: &[CharacterPosition],
     state: &State,
     tree: &mut tree::Tree,
 ) -> Vec<TextChunk> {
-    let mut chunks = Vec::new();
-    let mut chars_count = 0;
-    let mut chunk_bytes_count = 0;
-    for child in text_elem.descendants().filter(|n| n.is_text()) {
-        let ref parent = child.parent().unwrap();
+    let mut iter_state = IterState {
+        chars_count: 0,
+        chunk_bytes_count: 0,
+        split_chunk: false,
+        text_flow: TextFlow::Horizontal,
+        chunks: Vec::new(),
+    };
+
+    collect_text_chunks_impl(text_node, text_node, pos_list, state, tree, &mut iter_state);
+
+    iter_state.chunks
+}
+
+fn collect_text_chunks_impl(
+    text_node: &TextNode,
+    parent: &svgdom::Node,
+    pos_list: &[CharacterPosition],
+    state: &State,
+    tree: &mut tree::Tree,
+    iter_state: &mut IterState,
+) {
+    for child in parent.children() {
+        if child.is_element() {
+            if child.is_tag_name(EId::TextPath) {
+                if !parent.is_tag_name(EId::Text) {
+                    // `textPath` can be set only as a direct `text` element child.
+                    iter_state.chars_count += count_chars(&child);
+                    continue;
+                }
+
+                match resolve_text_flow(child.clone(), state) {
+                    Some(v) => {
+                        iter_state.text_flow = v;
+                    }
+                    None => {
+                        // Skip an invalid text path and all it's children.
+                        // We have to update the chars count,
+                        // because `pos_list` was calculated including this text path.
+                        iter_state.chars_count += count_chars(&child);
+                        continue;
+                    }
+                }
+
+                iter_state.split_chunk = true;
+            }
+
+            collect_text_chunks_impl(text_node, &child, pos_list, state, tree, iter_state);
+
+            iter_state.text_flow = TextFlow::Horizontal;
+
+            // Next char after `textPath` should be split too.
+            if child.is_tag_name(EId::TextPath) {
+                iter_state.split_chunk = true;
+            }
+
+            continue;
+        }
 
         if !style::is_visible_element(parent, state.opt) {
-            chars_count += child.text().chars().count();
+            iter_state.chars_count += child.text().chars().count();
             continue;
         }
 
@@ -190,7 +273,7 @@ pub fn collect_text_chunks(
         let font_size = units::resolve_font_size(parent, state);
         if !(font_size > 0.0) {
             // Skip this span.
-            chars_count += child.text().chars().count();
+            iter_state.chars_count += child.text().chars().count();
             continue;
         }
 
@@ -198,7 +281,7 @@ pub fn collect_text_chunks(
             Some(v) => v,
             None => {
                 // Skip this span.
-                chars_count += child.text().chars().count();
+                iter_state.chars_count += child.text().chars().count();
                 continue;
             }
         };
@@ -213,7 +296,7 @@ pub fn collect_text_chunks(
             stroke: style::resolve_stroke(parent, true, state, tree),
             font,
             font_size,
-            decoration: resolve_decoration(text_elem, parent, state, tree),
+            decoration: resolve_decoration(text_node, parent, state, tree),
             visibility: parent.find_enum(AId::Visibility),
             baseline_shift: resolve_baseline_shift(parent, state),
             letter_spacing,
@@ -227,37 +310,44 @@ pub fn collect_text_chunks(
             // Create a new chunk if:
             // - this is the first span (yes, position can be None)
             // - text character has an absolute coordinate assigned to it (via x/y attribute)
-            let is_new_chunk =    pos_list[chars_count].x.is_some()
-                || pos_list[chars_count].y.is_some()
-                || chunks.is_empty();
+            // - `c` is the first char of the `textPath`
+            // - `c` is the first char after `textPath`
+            let is_new_chunk =
+                pos_list[iter_state.chars_count].x.is_some()
+                    || pos_list[iter_state.chars_count].y.is_some()
+                    || iter_state.split_chunk
+                    || iter_state.chunks.is_empty();
+
+            iter_state.split_chunk = false;
 
             if is_new_chunk {
-                chunk_bytes_count = 0;
+                iter_state.chunk_bytes_count = 0;
 
                 let mut span2 = span.clone();
                 span2.start = 0;
                 span2.end = char_len;
 
-                chunks.push(TextChunk {
-                    x: pos_list[chars_count].x,
-                    y: pos_list[chars_count].y,
+                iter_state.chunks.push(TextChunk {
+                    x: pos_list[iter_state.chars_count].x,
+                    y: pos_list[iter_state.chars_count].y,
                     anchor,
                     spans: vec![span2],
+                    text_flow: iter_state.text_flow.clone(),
                     text: c.to_string(),
                 });
             } else if is_new_span {
                 // Add this span to the last text chunk.
                 let mut span2 = span.clone();
-                span2.start = chunk_bytes_count;
-                span2.end = chunk_bytes_count + char_len;
+                span2.start = iter_state.chunk_bytes_count;
+                span2.end = iter_state.chunk_bytes_count + char_len;
 
-                if let Some(chunk) = chunks.last_mut() {
+                if let Some(chunk) = iter_state.chunks.last_mut() {
                     chunk.text.push(c);
                     chunk.spans.push(span2);
                 }
             } else {
                 // Extend the last span.
-                if let Some(chunk) = chunks.last_mut() {
+                if let Some(chunk) = iter_state.chunks.last_mut() {
                     chunk.text.push(c);
                     if let Some(span) = chunk.spans.last_mut() {
                         debug_assert_ne!(span.end, 0);
@@ -267,23 +357,57 @@ pub fn collect_text_chunks(
             }
 
             is_new_span = false;
-            chars_count += 1;
-            chunk_bytes_count += char_len;
+            iter_state.chars_count += 1;
+            iter_state.chunk_bytes_count += char_len;
         }
     }
+}
 
-    chunks
+fn resolve_text_flow(
+    node: svgdom::Node,
+    state: &State,
+) -> Option<TextFlow> {
+    let av = node.attributes().get_value(AId::Href).cloned();
+    let path_node = match av {
+        Some(AValue::Link(n)) => n.clone(),
+        _ => return None,
+    };
+
+    if !path_node.is_tag_name(EId::Path) {
+        return None;
+    }
+
+    let segments = match path_node.attributes().get_value(AId::D).cloned() {
+        Some(AValue::Path(path)) => crate::convert::path::convert(path),
+        _ => return None,
+    };
+
+    if segments.len() < 2 {
+        return None;
+    }
+
+    let start_offset = node.attributes().get_length_or(AId::StartOffset, Length::zero());
+    let start_offset = if start_offset.unit == Unit::Percent {
+        // 'If a percentage is given, then the `startOffset` represents
+        // a percentage distance along the entire path.'
+        let path_len = crate::utils::path_length(&segments);
+        path_len * (start_offset.num / 100.0)
+    } else {
+        node.resolve_length(AId::StartOffset, state, 0.0)
+    };
+
+    Some(TextFlow::Path(Rc::new(TextPath {
+        start_offset,
+        segments,
+    })))
 }
 
 pub fn resolve_rendering_mode(
-    node: &svgdom::Node,
+    text_node: &TextNode,
     state: &State,
 ) -> tree::ShapeRendering {
-    // `text-rendering` can be set only on a `text` element.
-    debug_assert!(node.is_tag_name(EId::Text));
-
-    let mode: tree::TextRendering = node.try_find_enum(AId::TextRendering)
-                                        .unwrap_or(state.opt.text_rendering);
+    let mode: tree::TextRendering = text_node.try_find_enum(AId::TextRendering)
+        .unwrap_or(state.opt.text_rendering);
 
     match mode {
         tree::TextRendering::OptimizeSpeed => tree::ShapeRendering::CrispEdges,
@@ -505,11 +629,11 @@ pub struct CharacterPosition {
 ///
 /// The result should be: `[100, 50, 120, None]`
 pub fn resolve_positions_list(
-    text_elem: &svgdom::Node,
+    text_node: &TextNode,
     state: &State,
 ) -> Vec<CharacterPosition> {
     // Allocate a list that has all characters positions set to `None`.
-    let total_chars = count_chars(text_elem);
+    let total_chars = count_chars(text_node);
     let mut list = vec![CharacterPosition {
         x: None,
         y: None,
@@ -518,7 +642,7 @@ pub fn resolve_positions_list(
     }; total_chars];
 
     let mut offset = 0;
-    for child in text_elem.descendants() {
+    for child in text_node.descendants() {
         if child.is_element() {
             let child_chars = count_chars(&child);
             macro_rules! push_list {
@@ -556,13 +680,13 @@ pub fn resolve_positions_list(
 ///
 /// Note: this algorithm differs from the position resolving one.
 pub fn resolve_rotate_list(
-    text_elem: &svgdom::Node,
+    text_node: &TextNode,
 ) -> Vec<f64> {
     // Allocate a list that has all characters angles set to `0.0`.
-    let mut list = vec![0.0; count_chars(text_elem)];
+    let mut list = vec![0.0; count_chars(text_node)];
     let mut last = 0.0;
     let mut offset = 0;
-    for child in text_elem.descendants() {
+    for child in text_node.descendants() {
         if child.is_element() {
             if let Some(num_list) = child.attributes().get_number_list(AId::Rotate) {
                 for i in 0..count_chars(&child) {
@@ -602,21 +726,21 @@ pub struct TextDecoration {
 ///
 /// `text` and `tspan` can point to the same node.
 fn resolve_decoration(
-    text: &svgdom::Node,
+    text_node: &TextNode,
     tspan: &svgdom::Node,
     state: &State,
     tree: &mut tree::Tree,
 ) -> TextDecoration {
     // TODO: explain the algorithm
 
-    let text_dec = conv_text_decoration(text);
+    let text_dec = conv_text_decoration(text_node);
     let tspan_dec = conv_text_decoration2(tspan);
 
     let mut gen_style = |in_tspan: bool, in_text: bool| {
         let n = if in_tspan {
             tspan.clone()
         } else if in_text {
-            text.clone()
+            (*text_node).clone()
         } else {
             return None;
         };
@@ -641,17 +765,15 @@ struct TextDecorationTypes {
 }
 
 /// Resolves the `text` node's `text-decoration` property.
-fn conv_text_decoration(node: &svgdom::Node) -> TextDecorationTypes {
-    debug_assert!(node.is_tag_name(EId::Text));
-
+fn conv_text_decoration(text_node: &TextNode) -> TextDecorationTypes {
     fn find_decoration(node: &svgdom::Node, value: &str) -> bool {
         node.ancestors().any(|n| n.attributes().get_str(AId::TextDecoration) == Some(value))
     }
 
     TextDecorationTypes {
-        has_underline: find_decoration(node, "underline"),
-        has_overline: find_decoration(node, "overline"),
-        has_line_through: find_decoration(node, "line-through"),
+        has_underline: find_decoration(text_node, "underline"),
+        has_overline: find_decoration(text_node, "overline"),
+        has_line_through: find_decoration(text_node, "line-through"),
     }
 }
 
@@ -747,12 +869,9 @@ fn resolve_font_weight(
 }
 
 fn count_chars(node: &svgdom::Node) -> usize {
-    let mut total = 0;
-    for child in node.descendants().filter(|n| n.is_text()) {
-        total += child.text().chars().count();
-    }
-
-    total
+    node.descendants()
+        .filter(|n| n.is_text())
+        .fold(0, |w, n| w + n.text().chars().count())
 }
 
 /// Converts the writing mode.
@@ -771,11 +890,8 @@ fn count_chars(node: &svgdom::Node) -> usize {
 /// it should affect the rendering.
 ///
 /// [SVG 2.0]: https://www.w3.org/TR/SVG2/text.html#WritingModeProperty
-pub fn convert_writing_mode(node: &svgdom::Node) -> WritingMode {
-    // `writing-mode` can be set only on a `text` element.
-    debug_assert!(node.is_tag_name(EId::Text));
-
-    if let Some(n) = node.find_node_with_attribute(AId::WritingMode) {
+pub fn convert_writing_mode(text_node: &TextNode) -> WritingMode {
+    if let Some(n) = text_node.find_node_with_attribute(AId::WritingMode) {
         match n.attributes().get_str_or(AId::WritingMode, "lr-tb") {
             "tb" | "tb-rl" => WritingMode::TopToBottom,
             _ => WritingMode::LeftToRight,